Organize source into folders and split large files
Group the flat EdmundCore sources into Model/Parsing/TextView/Editing/
Rendering folders and edmd into App/Views/Settings (SPM finds sources
recursively, so Package.swift is unchanged; all moves are git renames).

Split the three largest files into focused extensions, no behavior change:
- EditorTextView.swift (749) -> +EditFlow, +SelectionTracking,
  +TypewriterScroll (core now 397)
- EditorTextView+Rendering.swift (722) -> +TableRendering (styleTableSpan)
  and +ListMarkerRendering (styleListItemSpan) (now 523)
- SyntaxHighlighter+Walker.swift (567) -> +WalkerInline for the inline
  emphasis visitors (now 384)

Document the edmd backronym at the top of main.swift.
