+++ Sources/EdmundCore/Editing/EditorTextView+ListContinuation.swift
