+++ Sources/EdmundCore/Model/Block.swift
