+++ Sources/EdmundCore/Model/Callout.swift
