+++ Sources/EdmundCore/Model/LineEnding.swift
