+++ Sources/EdmundCore/Model/ListIndentState.swift
