+++ Sources/EdmundCore/Model/StatusBarPrefs.swift
