+++ Sources/EdmundCore/Rendering/EditorTextView+ImageRendering.swift
