+++ Sources/EdmundCore/Rendering/EditorTextView+TableSupport.swift
