+++ Sources/EdmundCore/TextView/EditorTextView+LazyStyling.swift
