+++ Sources/EdmundCore/TextView/EditorTextView+TextKit2.swift
