+++ Sources/edmd/Views/StatusBarView.swift
